// web/src/Definitions/Interfaces/connectionContext.interfaces.tsx
import { AxiosRequestConfig, AxiosResponseHeaders, RawAxiosResponseHeaders } from 'axios';

export interface Response<T> {
  data: T;
  statusText: string;
  status?: number;
  headers: RawAxiosResponseHeaders | AxiosResponseHeaders;
  config?: AxiosRequestConfig;
  request?: any;
  message: string;
}

export type ConnectionProps = {
  post<T = any, R = Response<T>>(path: string, data?: any, config?: AxiosRequestConfig): Promise<R>;
  get<T = any, R = Response<T>>(path: string, config?: AxiosRequestConfig): Promise<R>;
  delete<T = any, R = Response<T>>(path: string, config?: AxiosRequestConfig): Promise<R>;
  put<T = any, R = Response<T>>(path: string, data?: any, config?: AxiosRequestConfig): Promise<R>;
  patch<T = any, R = Response<T>>(
    path: string,
    data?: any,
    config?: AxiosRequestConfig
  ): Promise<R>;
  updateToken: (token?: string) => void;
  token?: string;
};